Fix import resolver
